Make process to be killed correctly

+++ archive_download_service/utils/process.py
@@ -3,11 +3,14 @@ import psutil
 
 def kill_process_tree(pid: int, including_parent: bool = True) -> None:
     timeout = 5
-    parent = psutil.Process(pid)
+    try:
+        parent = psutil.Process(pid)
+    except psutil.NoSuchProcess:
+        return
     children = parent.children(recursive=True)
     for child in children:
         child.kill()
     psutil.wait_procs(children, timeout=timeout)
-    if including_parent:
+    if parent.is_running() and including_parent:
         parent.kill()
         parent.wait(timeout)
